requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
openai
huggingface_hub